chore: rename to migrate
