+++ crates/cli/commands/src/db/mod.rs
@@ -16,7 +16,6 @@ mod diff;
 mod get;
 mod list;
 mod migrate;
-mod mdbx_to_mdbx;
 mod repair_trie;
 mod settings;
 mod static_file_header;
@@ -57,8 +56,6 @@ pub enum Subcommands {
     Clear(clear::Command),
     /// Migrate database to another location
     Migrate(migrate::Command),
-    /// Copy database to another location (mdbx-to-mdbx)
-    MdbxToMdbx(mdbx_to_mdbx::Command),
     /// Verifies trie consistency and outputs any inconsistencies
     RepairTrie(repair_trie::Command),
     /// Reads and displays the static file segment header
@@ -79,7 +76,7 @@ macro_rules! db_exec {
     ($env:expr, $tool:ident, $N:ident, $access_rights:expr, $command:block) => {
         let Environment { provider_factory, .. } = $env.init::<$N>($access_rights)?;
 
-        let $tool = DbTool::new(provider_factory)?;
+        let $tool = DbTool::new(provider_factory.clone())?;
         $command;
     };
 }
@@ -165,13 +162,6 @@ impl<C: ChainSpecParser<ChainSpec: EthChainSpec + EthereumHardforks>> Command<C>
                 });
             }
             Subcommands::Migrate(command) => {
-                // Get database arguments from system configuration
-                let db_args = self.env.db.database_args();
-                db_ro_exec!(self.env, tool, N, {
-                    command.execute(tool.provider_factory.db_ref(), &db_args)?;
-                });
-            }
-            Subcommands::MdbxToMdbx(command) => {
                 // Get database arguments from system configuration
                 let db_args = self.env.db.database_args();
                 db_ro_exec!(self.env, tool, N, {
